+++ projects/ngx-stripe-docs/src/app/examples/address-element.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, OnInit, ViewChild, inject } from '@angular/core';
+import { AfterViewInit, ChangeDetectorRef, Component, OnInit, ViewChild, inject } from '@angular/core';
 import { ReactiveFormsModule, UntypedFormBuilder, Validators } from '@angular/forms';
 
 import {
@@ -47,6 +47,7 @@ export default class AddressElementExampleComponent implements OnInit, AfterView
 
   private readonly fb = inject(UntypedFormBuilder);
   private readonly plutoService = inject(NgStrPlutoService);
+  private readonly cdr = inject(ChangeDetectorRef);
 
   stripeTest = this.fb.group({
     name: ['Angular v12', [Validators.required]],
@@ -71,6 +72,7 @@ export default class AddressElementExampleComponent implements OnInit, AfterView
       })
       .subscribe((pi) => {
         this.elementsOptions.clientSecret = pi.client_secret;
+        this.cdr.detectChanges();
       });
   }